+++ components/connection/impl/src/main/java/com/flipperdevices/connection/impl/viewmodel/ConnectionTabStateMapper.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.connection.impl.viewmodel
 
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import com.flipperdevices.bottombar.model.TabState
 import com.flipperdevices.connection.impl.R
@@ -18,6 +19,8 @@ object ConnectionTabStateMapper {
                 notSelectedIcon = R.drawable.ic_no_device,
                 text = stringResource(id = R.string.connection_status_no_device),
                 selectedColor = LocalPallet.current.bottomBarSelectedFlipperStatus,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected
             )
             ConnectionStatusState.Disconnected -> TabState.Static(
@@ -25,6 +28,8 @@ object ConnectionTabStateMapper {
                 notSelectedIcon = R.drawable.ic_disconnected,
                 text = stringResource(id = R.string.connection_status_not_connected),
                 selectedColor = LocalPallet.current.bottomBarSelectedFlipperStatus,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected
             )
             ConnectionStatusState.Connecting -> TabState.Animated(
@@ -34,6 +39,8 @@ object ConnectionTabStateMapper {
                 notSelectedBackground = R.drawable.ic_connection_unselected_bg,
                 text = stringResource(id = R.string.connection_status_connecting),
                 selectedColor = LocalPallet.current.bottomBarSelectedFlipperStatus,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected,
                 textDotsAnimated = true
             )
@@ -42,16 +49,18 @@ object ConnectionTabStateMapper {
                 notSelectedIcon = R.drawable.ic_no_device,
                 text = stringResource(id = R.string.connection_status_unsupported),
                 selectedColor = LocalPallet.current.bottomBarUnsupported,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected,
-                unselectedColorIcon = LocalPallet.current.bottomBarUnsupported
             )
             ConnectionStatusState.Connected -> TabState.Static(
                 selectedIcon = R.drawable.ic_connected_filled,
                 notSelectedIcon = R.drawable.ic_connected,
                 text = stringResource(id = R.string.connection_status_connected),
                 selectedColor = LocalPallet.current.accentSecond,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected,
-                unselectedColorIcon = LocalPallet.current.accentSecond
             )
             is ConnectionStatusState.Synchronization -> TabState.Animated(
                 selectedIcon = R.drawable.ic_syncing_selected,
@@ -64,6 +73,8 @@ object ConnectionTabStateMapper {
                     connectionState.progress.roundPercentToString()
                 ),
                 selectedColor = LocalPallet.current.accentSecond,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected
             )
             is ConnectionStatusState.Synchronized -> TabState.Static(
@@ -71,8 +82,9 @@ object ConnectionTabStateMapper {
                 notSelectedIcon = R.drawable.ic_synced,
                 text = stringResource(id = R.string.connection_status_synced),
                 selectedColor = LocalPallet.current.accentSecond,
+                selectedColorIcon = Color.Unspecified,
+                unselectedColorIcon = Color.Unspecified,
                 unselectedColor = LocalPallet.current.bottomBarUnselected,
-                unselectedColorIcon = LocalPallet.current.accentSecond
             )
         }
     }