+++ components/nfc/attack/api/src/main/java/com/flipperdevices/nfc/attack/api/NfcAttackApi.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.nfc.attack.api
-
-import kotlinx.coroutines.flow.Flow
-
-interface NfcAttackApi {
-    fun notificationCount(): Flow<Int>
-}