+++ components/toolstab/impl/src/main/java/com/flipperdevices/toolstab/impl/composable/elements/MifareClassicComposable.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.nfc.attack.impl.composable.elements
+package com.flipperdevices.toolstab.impl.composable.elements
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -14,7 +14,6 @@ import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -23,8 +22,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.core.ui.theme.LocalPalletV2
 import com.flipperdevices.core.ui.theme.LocalTypography
-import com.flipperdevices.nfc.attack.impl.R
+import com.flipperdevices.toolstab.impl.R
 import com.flipperdevices.core.ui.res.R as DesignSystem
 
 @Composable
@@ -109,15 +109,11 @@ private fun MifareClassicMfKey32(
             modifier = Modifier
                 .size(16.dp)
                 .clip(CircleShape)
-                .background(LocalPallet.current.accent),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "!",
-                style = LocalTypography.current.monoSpaceM10,
-                color = LocalPallet.current.onFlipperButton
-            )
-        }
+                .background(LocalPalletV2.current.action.blackAndWhite.border.whiteOnColor)
+                .padding(1.dp)
+                .clip(CircleShape)
+                .background(LocalPalletV2.current.action.fwUpdate.background.primary.default),
+        )
     }
 
     Icon(